+++ resolwe_bio/tests/processes/test_support_processors.py
@@ -548,3 +548,23 @@ class SupportProcessorTestCase(KBBioProcessTestCase):
 
         dedup_paired = self.run_process('umi-tools-dedup', {'alignment': bam_paired.id})
         self.assertFile(dedup_paired, 'stats', './corall/output/dedup_paired_stats.txt')
+
+    @tag_process('seqtk-rev-complement-single', 'seqtk-rev-complement-paired')
+    def test_reverse_complement(self):
+        with self.preparation_stage():
+            single_reads = self.prepare_reads(['hs_slamseq_R1.fastq.gz'])
+
+            paired_reads = self.prepare_paired_reads(['hs_slamseq_R1.fastq.gz'], ['hs_slamseq_R2.fastq.gz'])
+
+        revcomp_single = self.run_process('seqtk-rev-complement-single', {'reads': single_reads.id})
+
+        self.assertFiles(revcomp_single, 'fastq', ['hs_slamseq_R1_complemented.fastq.gz'],
+                            compression='gzip')
+
+        revcomp_paired = self.run_process('seqtk-rev-complement-paired', {
+            'reads': paired_reads.id,
+            'select_mate': 'Mate 1'
+        })
+
+        self.assertFiles(revcomp_paired, 'fastq', ['hs_slamseq_R1_complemented.fastq.gz'],compression='gzip')
+        self.assertFiles(revcomp_paired, 'fastq2', ['hs_slamseq_R2.fastq.gz'],compression='gzip')